remove abstractions
We abstracted to quickly and I'm not convinced the abstractions we've made are good ones.  This commit removes unneccessary abstractions and we can revisit what the right abstractions are once we begin writing more verbose specs.

- removed `ActiveInteractor::Context::AttributeAssignment`
- removed `ActiveInteractor::Context::AttributeRegistration`
- removed `ActiveInteractor::Interactor::Callbacks`
- removed `ActiveInteractor::Interactor::ContextMethods`
- removed `ActiveInteractor::Interactor::InteractionMethods`
- removed `ActiveInteractor::Interactor::OptionMethods`
