+++ .rubocop/metrics.yml
@@ -1,3 +1,5 @@
 Metrics/BlockLength:
   Exclude:
     - '*.gemspec'
+Metrics/ClassLength:
+  Enabled: false
