+++ lib/active_interactor/context/attribute.rb
@@ -2,37 +2,10 @@
 
 module ActiveInteractor
   module Context
-    # An attribute of a {ActiveInteractor::Context::Base context}
-    #
-    # @api private
-    #
-    # @!attribute [r] description
-    #  @return [String, nil] the description of the attribute
-    #
-    # @!attribute [r] error_messages
-    #  @return [Array<Symbol>] the error messages for the attribute
-    #
-    # @!attribute [r] name
-    #  @return [Symbol] the name of the attribute
     class Attribute
-      # The default value when no default is provided
-      #
-      # @visibility private
       NO_DEFAULT_VALUE = :__no_default_value__
       attr_reader :description, :error_messages, :name
 
-      # Create a new instance of {ActiveInteractor::Context::Attribute}
-      #
-      # @param owner [ActiveInteractor::Context::Base] the {ActiveInteractor::Context::Base context} object the
-      #  attribute belongs to.
-      # @param [Object] name
-      # @param [Object] type
-      # @param [Object] description
-      # @param [Hash{Symbol => Object}] options
-      # @option options [Boolean] :required (false) whether or not the attribute is required
-      # @option options [Object] :default (NO_DEFAULT_VALUE) the default value for the attribute
-      #
-      # @return [ActiveInteractor::Context::Attribute] the new instance of
       def initialize(owner, name, type, description = nil, **options)
         @owner = owner
         @name = name.to_sym
@@ -42,49 +15,29 @@ module ActiveInteractor
         @error_messages = []
       end
 
-      # Assign a value to the attribute
-      #
-      # @param value [Object] the value to assign to the attribute
-      #
-      # @return [Object] the value assigned to the attribute
       def assign_value(value)
         @user_provided_value = value
       end
 
-      # The default value for the attribute
-      #
-      # @return [Object, nil] the default value for the attribute
       def default_value
         return if @options[:default] == NO_DEFAULT_VALUE
 
         @options[:default]
       end
 
-      # Whether or not the attribute is required
-      #
-      # @return [Boolean] whether or not the attribute is required
       def required?
         @options[:required]
       end
 
-      # The type of the attribute
-      #
-      # @return [Object] the type of the attribute
       def type
         @type_expression
       end
 
-      # Validate the attribute
-      #
-      # @return [void]
       def validate!
         validate_presence!
         validate_type!
       end
 
-      # The value of the attribute
-      #
-      # @return [Object] the value of the attribute
       def value
         @user_provided_value || default_value
       end
