+++ lib/active_interactor/context/attribute_validation.rb
@@ -2,7 +2,6 @@
 
 module ActiveInteractor
   module Context
-    # Methods for Attribute Validation
     module AttributeValidation
       extend ActiveSupport::Concern
 
@@ -12,9 +11,6 @@ module ActiveInteractor
 
       protected
 
-      # call {ActiveInteractor::Context::Attribute#validate!} on each attribute
-      #
-      # @visibility private
       def validate_attributes!
         attribute_set.attributes.each do |attribute|
           attribute.validate!
