+++ spec/active_interactor/interactor_spec.rb
@@ -7,22 +7,6 @@ RSpec.describe ActiveInteractor::Interactor do
     expect(described_class::Base).to be_a Class
   end
 
-  it 'is expected to have ::Callbacks' do
-    expect(described_class::Callbacks).to be_a Module
-  end
-
-  it 'is expected to have ::ContextMethods' do
-    expect(described_class::ContextMethods).to be_a Module
-  end
-
-  it 'is expected to have ::InteractionMethods' do
-    expect(described_class::InteractionMethods).to be_a Module
-  end
-
-  it 'is expected to have ::OptionMethods' do
-    expect(described_class::OptionMethods).to be_a Module
-  end
-
   it 'is expected to have ::Options' do
     expect(described_class::Options).to be_a Class
   end
